--- src/main/java/webserver/handler/Handler.java
@@ -1,12 +1,0 @@
-package webserver.handler;
-
-import webserver.request.HttpRequest;
-import webserver.response.HttpResponse;
-
-import java.io.IOException;
-
-@FunctionalInterface
-public interface Handler {
-
-    HttpResponse handle(HttpRequest request) throws IOException;
-}
