+++ src/main/java/webserver/http/HttpMethod.java
@@ -1,0 +1,7 @@
+package webserver.http;
+
+public enum HttpMethod {
+
+    GET,
+    POST;
+}
