+++ src/main/java/webserver/mvc/Handler.java
@@ -1,0 +1,12 @@
+package webserver.mvc;
+
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
+
+import java.io.IOException;
+
+@FunctionalInterface
+public interface Handler {
+
+    HttpResponse handle(HttpRequest request) throws IOException;
+}
