--- src/main/java/webserver/request/HttpMethod.java
@@ -1,7 +1,0 @@
-package webserver.request;
-
-public enum HttpMethod {
-
-    GET,
-    POST;
-}
