--- src/main/java/webserver/response/FileView.java
@@ -1,28 +1,0 @@
-package webserver.response;
-
-import utils.FileIoUtils;
-
-import java.util.Optional;
-
-public class FileView {
-
-    private final String basePath;
-
-    public FileView(String basePath) {
-        this.basePath = basePath;
-    }
-
-
-    public FileView() {
-        this("");
-    }
-
-    public Optional<byte[]> readFromPath(String path) {
-        try {
-            return Optional.of(FileIoUtils.loadFileFromClasspath(basePath + path));
-        } catch (Exception ignored) {
-        }
-        return Optional.empty();
-    }
-
-}
