--- src/main/java/webserver/response/HttpStatus.java
@@ -1,24 +1,0 @@
-package webserver.response;
-
-public enum HttpStatus {
-    OK(200, "OK"),
-    FOUND(302, "Found"),
-    NOT_FOUND(404, "Not Found");
-
-    private final int code;
-    private final String reasonPhrase;
-
-    HttpStatus(int code, String reasonPhrase) {
-        this.code = code;
-        this.reasonPhrase = reasonPhrase;
-    }
-
-    public int getCode() {
-        return code;
-    }
-
-    public String getReasonPhrase() {
-        return reasonPhrase;
-    }
-
-}
